Fixed pom

+++ src/main/java/splugins/test_grow_3D.java
@@ -78,7 +78,7 @@ public class test_grow_3D implements PlugIn {
 		Utility3D my3D=new Utility3D();
 		//short [] arr=Utility3D.ip_to_array(img);
 		//my3D.grow_until_neighbor(arr,  img.getWidth(),  img.getHeight(), img.getNSlices());
-		ArrayList <ArrayList <int []>> mylist=my3D.find_blobs((byte []) img.getProcessor().getPixels(),  img.getWidth(), img.getHeight(),  img.getNSlices());
+		ArrayList <ArrayList <int []>> mylist=my3D.find_blobs((short []) img.getProcessor().getPixels(),  img.getWidth(), img.getHeight(),  img.getNSlices());
 		short [] arr=my3D.blobarray_to_imgarray(mylist, width, height, slices);
 		ImagePlus new_img=Utility3D.imgarray_to_ip(arr,  img.getWidth(),  img.getHeight(),  img.getNSlices());
 		new_img.show();
